+++ src/view/filter.js
@@ -1,8 +1,6 @@
 import Abstract from "./abstract";
 
-const createFilterItemTemplate = (filter, isChecked) => {
-  const {name, count} = filter;
-
+const createFilterItemTemplate = ({name, count}, isChecked) => {
   return (
     `<input
       type="radio"
