+++ src/view/tasks.js
@@ -2,7 +2,7 @@ import Abstract from "./abstract";
 
 const createTasksTemplate = () => `<div class='board__tasks'></div>`;
 
-export default class Tasks extends Abstract{
+export default class Tasks extends Abstract {
   getTemplate() {
     return createTasksTemplate();
   }
